Move SessionsComparison component to containers directory
