+++ src/containers/SessionsComparison/SessionsComparison.css
