+++ src/containers/SessionsComparison/SessionsComparison.scss
